Remove factory pattern and remove some related duplication
