--- src/browser.js
@@ -1,4 +1,0 @@
-/* eslint-env commonjs */
-import m from "./index.js"
-if (typeof module !== "undefined") module["exports"] = m
-else window.m = m
