--- src/promise/promise.js
@@ -1,24 +1,0 @@
-/* eslint-env commonjs */
-import PromisePolyfill from "./polyfill.js"
-
-let exportedPromise
-
-if (typeof window !== "undefined") {
-	if (typeof window.Promise === "undefined") {
-		window.Promise = PromisePolyfill
-	} else if (!window.Promise.prototype.finally) {
-		window.Promise.prototype.finally = PromisePolyfill.prototype.finally
-	}
-	exportedPromise = window.Promise
-} else if (typeof global !== "undefined") {
-	if (typeof global.Promise === "undefined") {
-		global.Promise = PromisePolyfill
-	} else if (!global.Promise.prototype.finally) {
-		global.Promise.prototype.finally = PromisePolyfill.prototype.finally
-	}
-	exportedPromise = global.Promise
-} else {
-	exportedPromise = PromisePolyfill
-}
-
-export default exportedPromise
