--- src/render/trust.js
@@ -1,6 +1,0 @@
-import Vnode from "./vnode"
-
-export default function(html) {
-	if (html == null) html = ""
-	return Vnode("<", undefined, undefined, html, undefined, undefined)
-}
