--- src/util/assign.js
@@ -1,8 +1,0 @@
-// This exists so I'm only saving it once.
-import hasOwn from "./hasOwn"
-
-export default Object.assign || function(target, source) {
-	for (var key in source) {
-		if (hasOwn.call(source, key)) target[key] = source[key]
-	}
-}
