--- src/util/hasOwn.js
@@ -1,2 +1,0 @@
-// This exists so I'm only saving it once.
-export default {}.hasOwnProperty
