+++ tests/promise/test.html
@@ -11,7 +11,7 @@
 		<script>
 			window.module = {exports:null}
 		</script>
-		<script src="../../promise/polyfill.js"></script>
+		<script src="../../temp/index.js"></script>
 		<script>
 			PromisePolyfill = module.exports
 			delete module.exports
